Add files via upload
PokemonClassifierGenOne

*HOW TO INSTALL*
Choose the appropriate installer for your hardware and install Python 3.7.0. Now that you have done that, download this program.
Open terminal/command prompt and simply copy-paste the location into terminal/command prompt. Once this is done, press enter/return and the program should be fully functional.

*HOW TO USE*
Once the program is open, you will be greeted with a question, "Which Pokemon would you like to learn about?". You can type in the name of any GENERATION ONE (Kanto) Pokemon and you will be given it's name, what kind of Pokemon it is (for example, Pikachu is the mouse Pokemon), its National Pokedex number, and it's typing.
If you would like to exit the program, you can type quit, quit(), end, or stop. This will terminate the loop that runs the question and thus close the program.